+++ spring-boot-demo-https/src/main/java/com/xkcoding/https/SpringBootDemoHttpsApplication.java
@@ -1,0 +1,21 @@
+package com.xkcoding.https;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * <p>
+ * 启动类
+ * </p>
+ *
+ * @author Chen.Chao
+ * @date Created in 2020/1/12 10:31
+ */
+@SpringBootApplication
+public class SpringBootDemoHttpsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringBootDemoHttpsApplication.class, args);
+    }
+
+}
