--- spring-boot-demo-https/src/main/java/com/xkcoding/springbootdemohttps/SpringBootDemoHttpsApplication.java
@@ -1,62 +1,0 @@
-package com.xkcoding.springbootdemohttps;
-
-import org.apache.catalina.Context;
-import org.apache.catalina.connector.Connector;
-import org.apache.tomcat.util.descriptor.web.SecurityCollection;
-import org.apache.tomcat.util.descriptor.web.SecurityConstraint;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
-import org.springframework.context.annotation.Bean;
-
-
-/**
- * <p>
- * SpringBoot启动类
- * </p>
- *
- * @package: com.xkcoding.https
- * @description: SpringBoot启动类
- * @author: Chen.Chao
- * @date 2020.01.12 10:31 am
- * @copyright: Copyright (c)
- * @version: V1.0
- * @modified: Chen.Chao
- */
-@SpringBootApplication
-public class SpringBootDemoHttpsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringBootDemoHttpsApplication.class, args);
-    }
-
-
-    @Bean
-    public Connector connector(){
-        Connector connector=new Connector("org.apache.coyote.http11.Http11NioProtocol");
-        connector.setScheme("http");
-        connector.setPort(80);
-        connector.setSecure(false);
-        connector.setRedirectPort(443);
-        return connector;
-    }
-
-    @Bean
-    public TomcatServletWebServerFactory tomcatServletWebServerFactory(Connector connector){
-        TomcatServletWebServerFactory tomcat=new TomcatServletWebServerFactory(){
-            @Override
-            protected void postProcessContext(Context context) {
-                SecurityConstraint securityConstraint=new SecurityConstraint();
-                securityConstraint.setUserConstraint("CONFIDENTIAL");
-                SecurityCollection collection=new SecurityCollection();
-                collection.addPattern("/*");
-                securityConstraint.addCollection(collection);
-                context.addConstraint(securityConstraint);
-            }
-        };
-        tomcat.addAdditionalTomcatConnectors(connector);
-        return tomcat;
-    }
-
-
-}
