+++ spring-boot-demo-https/src/main/resources/application.yml
@@ -1,7 +1,7 @@
 server:
   ssl:
     # 证书路径
-    key-store: spring-boot-demo-https\src\main\resources\server.keystore
+    key-store: classpath:server.keystore
     key-alias: tomcat
     enabled: true
     key-store-type: JKS
@@ -9,4 +9,3 @@ server:
     key-store-password: 123456
     # 浏览器默认端口 和 80 类似
   port: 443
-#debug: true
