+++ spring-boot-demo-https/src/test/java/com/xkcoding/https/SpringBootDemoHttpsApplicationTests.java
@@ -1,4 +1,4 @@
-package com.xkcoding.springbootdemohttps;
+package com.xkcoding.https;
 
 import org.junit.Test;
 import org.springframework.boot.test.context.SpringBootTest;
